Can select annotation
* Created abstract class PlaceablePetriNetComponent which has getters
  for x, y, height and width
* Select annotations in petrinet controller

+++ pipe-core/src/main/java/pipe/petrinet/unfold/Expander.java
@@ -212,7 +212,7 @@ public class Expander {
         return arcWeights;
     }
 
-    private Place getNewPlace(Place original, Double newX, Double newY, int tokenCount, String id) {
+    private Place getNewPlace(Place original, int newX, int newY, int tokenCount, String id) {
         if (newPlaces.containsKey(id)) {
             return newPlaces.get(id);
         }
